Kez Zoo Griffin animating and map making
Sending in multiple pushes! Did way too much on this build to include it all but to summarize animated the griffin and created a basic flight system. Then messed around with my Kez Zoo map and added a lot of assets and blocking volumes. Most definitely not going to load all that fast and rip building lighting but once you get access to it it's fun to play with :D Please do not edit my zoo without chatting with me first but you are more that welcome to play in it!