Mount and Rider Both Spawning Now
Both the mount and rider are spawning in after the character select menu because I had to put the true player character models in the character select screen instead of the placeholders. But because of this, I still need to find a workaround because the Griffin model seems to override the camera in the beginning and trying to figure out how to stop it. Switch gamemodes and default spawns don't seem to work. Plus, when spawning after the character select, can seem to get them to spawn look straight ahead.